backend/requirements.txt
Flask
Flask-Cors
pandas
openpyxl
python-dotenv
openai